fix wildcard handling
JList is initialized with 0 or 1 arguments. The original is this:
        if (arguments.length === 1) {
            sequence.push(arguments[0]);
        }
If an argument is passed, it is added. The new code mimics this.

The second issue is in the wildard handling. In the list case, arrays
need to be flattended. The issue comes from the unnecessary handling of
maps. Like scalar types, those simply need to be added to the result